pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "market-rally"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "yfinance>=0.2.30",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "hmmlearn>=0.3",
    "lxml>=4.9",
    "joblib>=1.3",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
    "anthropic>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]
discord = [
    "discord.py>=2.3",
]
alpaca = [
    "alpaca-py>=0.29",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E402"]  # matplotlib.use("Agg") requires imports after call

[tool.ruff.lint.per-file-ignores]
"src/rally/live/main.py" = ["I001"]
"src/rally/backtest/optimize.py" = ["I001"]
"src/rally/backtest/universe_bt.py" = ["I001"]
